refactor(api): retire deprecated and dashboard-orphaned routes
Remove four /api dashboard routes with no frontend caller and no other
documented consumer (slop-removal 0908, task C6):

- POST /api/accounts/{id}/export and /export/opencode-auth: deprecated
  since 2026-07-12-unify-auth-export; POST /export/auth already returns
  both payloads. Drops AccountExportResponse,
  AccountOpenCodeAuthExportResponse and the two service methods.
- DELETE /api/sticky-sessions/{kind}/{key}: frontend client removed in
  #2183; the UI deletes single rows through POST /delete. Drops
  StickySessionDeleteResponse and StickySessionsService.delete_entry.
- GET /api/conversation-archive/files: the dashboard only calls
  /records. Drops ConversationArchiveFileResponse, ConversationArchiveFile,
  list_archive_files and _date_from_filename.

Kept: fleet/* (fleet-summary spec, Bearer operator API), usage/* and
audit-logs (undocumented but out of scope; follow-up decision), and
GET /api/conversations/{id} which the audit misattributed to
request-logs (documented, spec-governed, called by the dashboard).

Tests move to the surviving routes; the OpenCode export scenarios fold
into test_account_auth_export.py. Frontend drops the /export MSW mock
and the unused ConversationArchiveFileSchema. OpenSpec change
retire-uncalled-dashboard-routes records the consumer audit with
MODIFIED account-auth-export / sticky-session-operations deltas and an
ADDED unified-auth-export requirement retiring the predecessor routes.
