+++ app/modules/conversation_archive/api.py
@@ -9,7 +9,6 @@ from fastapi.concurrency import run_in_threadpool
 from app.core.auth.dependencies import require_dashboard_admin_access, set_dashboard_error_format
 from app.modules.conversation_archive import service
 from app.modules.conversation_archive.schemas import (
-    ConversationArchiveFileResponse,
     ConversationArchiveRecordResponse,
     ConversationArchiveRecordsResponse,
 )
@@ -19,21 +18,6 @@ router = APIRouter(
     tags=["dashboard"],
     dependencies=[Depends(require_dashboard_admin_access), Depends(set_dashboard_error_format)],
 )
-
-
-@router.get("/files", response_model=list[ConversationArchiveFileResponse])
-def list_conversation_archive_files() -> list[ConversationArchiveFileResponse]:
-    files = service.list_archive_files()
-    return [
-        ConversationArchiveFileResponse(
-            name=file.name,
-            date=file.date,
-            size_bytes=file.size_bytes,
-            compressed=file.compressed,
-            modified_at=file.modified_at,
-        )
-        for file in files
-    ]
 
 
 @router.get("/records", response_model=ConversationArchiveRecordsResponse)
