+++ frontend/src/features/conversation-archive/schemas.test.ts
@@ -1,23 +1,8 @@
 import { describe, expect, it } from "vitest";
 
-import {
-  ConversationArchiveFileSchema,
-  ConversationArchiveRecordsResponseSchema,
-} from "@/features/conversation-archive/schemas";
+import { ConversationArchiveRecordsResponseSchema } from "@/features/conversation-archive/schemas";
 
 describe("conversation archive schemas", () => {
-  it("parses archive file metadata", () => {
-    const parsed = ConversationArchiveFileSchema.parse({
-      name: "2026-04-29.jsonl.gz",
-      date: "2026-04-29",
-      sizeBytes: 1234,
-      compressed: true,
-      modifiedAt: "2026-04-29T10:00:00Z",
-    });
-
-    expect(parsed.compressed).toBe(true);
-  });
-
   it("parses records with arbitrary payloads", () => {
     const parsed = ConversationArchiveRecordsResponseSchema.parse({
       records: [
