+++ tests/integration/test_account_auth_export.py
@@ -1,15 +1,73 @@
 from __future__ import annotations
 
+import asyncio
+import base64
 import json
+from collections.abc import Mapping
 from typing import cast
 
 import pytest
+from sqlalchemy import select
 
 from app.core.auth import generate_unique_account_id
-
-from .test_account_opencode_auth_export import _make_auth_json, _wait_for_audit_log
+from app.db.models import AuditLog
+from app.db.session import SessionLocal
 
 pytestmark = pytest.mark.integration
+
+
+def _encode_jwt(payload: Mapping[str, object]) -> str:
+    raw = json.dumps(payload, separators=(",", ":")).encode("utf-8")
+    body = base64.urlsafe_b64encode(raw).rstrip(b"=").decode("ascii")
+    return f"header.{body}.sig"
+
+
+def _make_auth_json(account_id: str, email: str, *, access_exp: int = 2_000_000_000) -> dict[str, object]:
+    id_payload = {
+        "email": email,
+        "chatgpt_account_id": account_id,
+        "https://api.openai.com/auth": {"chatgpt_plan_type": "plus"},
+    }
+    access_payload = {
+        "exp": access_exp,
+        "https://api.openai.com/auth": {"chatgpt_account_id": account_id},
+    }
+    return {
+        "tokens": {
+            "idToken": _encode_jwt(id_payload),
+            "accessToken": _encode_jwt(access_payload),
+            "refreshToken": "refresh-token",
+            "accountId": account_id,
+        },
+    }
+
+
+def _make_auth_json_without_account_id(email: str, *, access_exp: int = 2_000_000_000) -> dict[str, object]:
+    id_payload = {
+        "email": email,
+        "https://api.openai.com/auth": {"chatgpt_plan_type": "plus"},
+    }
+    access_payload = {"exp": access_exp}
+    return {
+        "tokens": {
+            "idToken": _encode_jwt(id_payload),
+            "accessToken": _encode_jwt(access_payload),
+            "refreshToken": "refresh-token",
+        },
+    }
+
+
+async def _wait_for_audit_log(action: str, *, attempts: int = 20) -> AuditLog:
+    for _ in range(attempts):
+        async with SessionLocal() as session:
+            result = await session.execute(
+                select(AuditLog).where(AuditLog.action == action).order_by(AuditLog.id.desc())
+            )
+            row = result.scalars().first()
+            if row is not None:
+                return row
+        await asyncio.sleep(0.05)
+    raise AssertionError(f"audit log not written for action={action}")
 
 
 @pytest.mark.asyncio
@@ -82,3 +140,40 @@ async def test_export_account_auth_missing_account_returns_404(async_client) -> 
 
     assert response.status_code == 404
     assert response.json()["error"]["code"] == "account_not_found"
+
+
+@pytest.mark.asyncio
+async def test_export_account_auth_keeps_unknown_account_id_null(async_client) -> None:
+    email = "unknown-combined@example.com"
+
+    import_response = await async_client.post(
+        "/api/accounts/import",
+        files={
+            "auth_json": (
+                "auth.json",
+                json.dumps(_make_auth_json_without_account_id(email)),
+                "application/json",
+            ),
+        },
+    )
+    assert import_response.status_code == 200
+    imported_account_id = import_response.json()["accountId"]
+
+    export_response = await async_client.post(f"/api/accounts/{imported_account_id}/export/auth")
+
+    assert export_response.status_code == 200
+    payload = export_response.json()
+    assert payload["account"]["accountId"] == imported_account_id
+    assert payload["account"]["chatgptAccountId"] is None
+    assert payload["codexAuthJson"]["tokens"]["account_id"] is None
+    assert payload["opencodeAuthJson"]["openai"]["accountId"] is None
+
+
+@pytest.mark.asyncio
+async def test_deprecated_export_routes_are_retired(async_client) -> None:
+    # The SPA catch-all turns an unmatched POST into a 405 partial match; either
+    # way the route is not served and no credential material leaves the server.
+    for path in ("/api/accounts/any-account/export", "/api/accounts/any-account/export/opencode-auth"):
+        response = await async_client.post(path)
+        assert response.status_code in (404, 405), path
+        assert not {"authJson", "tokens", "codexAuthJson", "opencodeAuthJson"} & response.json().keys(), path
